changing setup for tutorials
There is now a editorLines property, which sets the height of the editor (defaults to 5). Additionally, the descriptions of individual steps are stored in a separate HTML file, which is referenced from the descfile property. These files should start with an html:div element, and then contain the content of the former desc property in the JSON file.