refactor(tutorials.js): simplify handleEditorChanges() function
This change moves the changeListener for the ACE editor from the handle EditorChanges() function
to the loadTutorialStep() function. Thus it avoids the necessity to stop change propagation and recursive call.
It also removes some unnecessary parameters from handleEditorChanges() as well as it simplifies the mechanism
to disallow/allow download and/or continuation with next step.