Update 104_rests_step-00-desc.html
Provides navigation to XML basics and Quickstart tutorials if the user is not yet able to complete the first task here. This adjustment also reinforces the sequence that the "100" numbering of the tutorials implies, so that no tutorial leaves the user feeling "stranded" if they land somewhere and don't know what to do next. Pedagogically, the goal is to minimize points of friction where the new learner might "abandon ship."